stats

+++ 0x09-static_libraries/_putchar.c
@@ -1,10 +1,11 @@
-#include "main.h"
-#include <stdio.h>
-#include <stdlib.h>
-#include <ctype.h>
-#include <string.h>
-
+#include <unistd.h>
+/**
+ * _putchar - writes the character c to stdout
+ * @c: The character to print
+ *
+ * Return: On success, the character written. On error, -1 is returned
+ */
 int _putchar(char c)
 {
-    return putchar(c);
+	return (putchar((int)c));
 }
